requirements.txt
django
psycopg2-binary
django-ckeditor
Pillow
